Added changes

+++ src/Nabeey.Service/DTOs/Quizzes/UserRatingDto.cs
@@ -4,7 +4,7 @@ namespace Nabeey.Domain.Entities.Quizzes;
 
 public class UserRatingDto
 {
-	public short Rating { get; set; }
-	public int Ball { get; set; }
+	public int Rating { get; set; }
+	public double Ball { get; set; }
 	public UserResultDto User { get; set; }
 }
